+++ packages/eslint-plugin-react-signals-hooks/tests/prefer-use-computed-in-react-component/simple.config.mjs
@@ -1,0 +1,32 @@
+import tsParser from '@typescript-eslint/parser';
+
+import reactSignalsHooksPlugin from '../../dist/esm/index.mjs';
+
+/** @type {import('eslint').Linter.Config[]} */
+export default [
+  {
+    ignores: ['node_modules/**', 'dist/**'],
+  },
+  {
+    files: ['*.tsx', '*.ts'],
+    plugins: {
+      'react-signals-hooks': reactSignalsHooksPlugin,
+    },
+    rules: {
+      // Rule under test
+      'react-signals-hooks/prefer-use-computed-in-react-component': 'warn',
+    },
+    languageOptions: {
+      parser: tsParser,
+      parserOptions: {
+        tsconfigRootDir: process.cwd(),
+        project: '../../tsconfig.tests.json',
+        ecmaVersion: 'latest',
+        sourceType: 'module',
+        ecmaFeatures: {
+          jsx: true,
+        },
+      },
+    },
+  },
+];