+++ packages/eslint-plugin-react-signals-hooks/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.0] - 2025-08-22
+
+### Added 2 new rules
+
+- `prefer-use-computed-in-react-component`
+  - Encourages using `useComputed(...)` inside React components when appropriate.
+  - Provides auto-fix to add or remove the `useComputed` import when necessary.
+
+- `prefer-use-signal-effect-in-react-component`
+  - Encourages using `useSignalEffect(...)` inside React components when appropriate.
+  - Provides auto-fix to add or remove the `useSignalEffect` import when necessary.
+
 ## [1.5.5] - 2025-08-19
 
 ### Fixed